requirements.txt
torch
torchvision
kaggle
Pillow
